requirements.txt
flask==3.1.0
flask-cors==5.0.1
supabase==2.11.0
python-dotenv==1.1.0
werkzeug==3.1.3
gunicorn==21.2.0
